requirements.txt
netmiko
pyvmomi==5.5.0
python-redmine
